Separated files into their on file
