+++ lib/navigation/route_manager.dart
@@ -1,53 +1,73 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_app/authenticate/log_in_page.dart';
 import 'package:flutter_app/authenticate/not_signed_in.dart';
-import 'package:flutter_app/model/user.dart';
 import 'package:flutter_app/screens/home/home_screen.dart';
-import 'package:flutter_app/screens/main_menu/add_event.dart';
-import 'package:flutter_app/screens/settings/profile.dart';
+import 'package:flutter_app/screens/main_menu/request_for_events.dart';
 import 'package:flutter_app/screens/settings/settings_page.dart';
-import 'package:provider/provider.dart';
 
 import '../main.dart';
-import '../size_config.dart';
 import '../wrapper.dart';
 
-Route<dynamic> generateRoute(RouteSettings settings){
-  MockUser authUser;
-  switch(settings.name){
+Route<dynamic> generateRoute(RouteSettings settings) {
+  switch (settings.name) {
     case RootRoute:
-      return MaterialPageRoute(builder: (context) {
-        return MyApp();
-      });
+      return MaterialPageRoute(
+        builder: (context) {
+          return MyApp();
+        },
+      );
+
     case HomeScreenRoute:
       return MaterialPageRoute(
-          builder: (context) => MainPage());
+        builder: (context) => MainPage(),
+      );
+
     case MenuOverviewRoute:
       return MaterialPageRoute(
-          builder: (context) => Wrapper(route: MenuOverviewRoute,));
+        builder: (context) => Wrapper(
+          route: MenuOverviewRoute,
+        ),
+      );
+
     case SettingsRoute:
       return MaterialPageRoute(
-        builder: (context) => SettingsPage());
+        builder: (context) => SettingsPage(),
+      );
+
     case ProfileRoute:
-      return MaterialPageRoute(builder: (context) {
-        return Wrapper(route: ProfileRoute,);
-      });
+      return MaterialPageRoute(
+        builder: (context) {
+          return Wrapper(
+            route: ProfileRoute,
+          );
+        },
+      );
+
     case EventRequestRoute:
       return MaterialPageRoute(
-          builder: (context) => RequestForEvents());
+        builder: (context) => RequestForEvents(),
+      );
+
     case SignInRoute:
       return MaterialPageRoute(
-          builder: (context) => Login());
+        builder: (context) => Login(),
+      );
+
     case NotSignedInRoute:
       return MaterialPageRoute(
-          builder: (context) => NotSignedIn());
+        builder: (context) => NotSignedIn(),
+      );
+
     case InitRoute:
       return MaterialPageRoute(
-          builder: (context) => InitScreen());
-    default:
-      return MaterialPageRoute(builder: (context) => MainPage());
-  }
+        builder: (context) => InitScreen(),
+      );
 
+    default:
+      return MaterialPageRoute(
+        builder: (context) => MainPage(),
+      );
+  }
 }
 
 const String InitRoute = "init";
