+++ lib/screens/main_menu/add_event.dart
@@ -5,8 +5,6 @@ import 'package:flutter/material.dart';
 import 'package:flutter_app/service/DatabaseService.dart';
 import 'package:flutter_app/size_config.dart';
 import 'package:flutter_app/widgets/category_grid.dart';
-import 'package:flutter_app/mixins/basic_mixin.dart';
-import 'package:flutter_app/widgets/event_request_widget.dart';
 import 'package:flutter_app/widgets/input_form_field.dart';
 import 'package:image_picker/image_picker.dart';
 import 'package:path/path.dart';
@@ -15,68 +13,6 @@ import 'package:datetime_picker_formfield/datetime_picker_formfield.dart';
 import '../../model/user.dart';
 import 'package:intl/intl.dart';
 import '../../theme.dart';
-import 'my_favorites.dart';
-import 'package:flutter_app/navigation/route_manager.dart' as router;
-
-class RequestForEvents extends StatefulWidget {
-  @override
-  _AddEventState createState() => _AddEventState();
-}
-
-class _AddEventState extends State<RequestForEvents> with BasicMixin {
-  List<EventRequest> requests = List();
-  List<Future<Event>> eventsFutures = List();
-  List<Future> fut = List();
-  double fadeTick = 0.3;
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: Text("Notifikationer"),
-      ),
-      body: body(context: context),
-    );
-  }
-
-  @override
-  Widget body({BuildContext context}) {
-    final authUser = Provider.of<MockUser>(context);
-    return StreamBuilder<List<EventRequest>>(
-      stream: DatabaseService().getEventRequests(authUser.uid),
-      builder: (context, snapshot) {
-        if (snapshot.hasData) {
-          for (var index in snapshot.data) {
-            eventsFutures
-                .add(DatabaseService().getEventFromUid(index.eventUid));
-          }
-        }
-        return FutureBuilder<List<Event>>(
-          future: Future.wait(eventsFutures),
-          builder: (context, AsyncSnapshot<List<Event>> snapshot) {
-            if (snapshot.hasData) {}
-            return snapshot.hasData
-                ? Column(
-                    children: [
-                      for (var ele in snapshot.data)
-                        FadeIn(
-                          fadeTick += 0.4,
-                          EventRequestWidget(
-                            eventTitle: ele.title,
-                          ),
-                        ),
-                    ],
-                  )
-                : Center(
-                    child: Container(
-                      child: Text("Ingen anmodninger"),
-                    ),
-                  );
-          },
-        );
-      },
-    );
-  }
-}
 
 class AddOrRepostEvent extends StatefulWidget {
   Event event;
@@ -347,137 +283,5 @@ class _AddOrRepostEventState extends State<AddOrRepostEvent> {
         ),
       ),
     );
-  }
-}
-
-class MenuOverview extends StatefulWidget {
-  @override
-  _MenuOverviewState createState() => _MenuOverviewState();
-}
-
-class _MenuOverviewState extends State<MenuOverview> with BasicMixin {
-  bool isFavorite = false;
-
-  @override
-  Widget appBar() {
-    return AppBar(
-        backgroundColor: Color.fromRGBO(30, 30, 60, 1),
-        centerTitle: true,
-        actions: [
-          Padding(
-            padding: EdgeInsets.only(right: 8),
-            child: GestureDetector(
-              onTap: () {
-                Navigator.pushNamed(this.context, router.EventRequestRoute);
-              },
-              child: Icon(Icons.notifications_none),
-            ),
-          ),
-        ],
-        title: Container(
-          width: SizeConfig.blockSizeHorizontal * 40,
-          child: Row(
-            mainAxisAlignment: MainAxisAlignment.spaceBetween,
-            children: [
-              Flexible(
-                flex: 1,
-                child: GestureDetector(
-                  onTap: () {
-                    setState(() {
-                      isFavorite = false;
-                    });
-                  },
-                  child: Container(
-                    width: double.infinity,
-                    decoration: BoxDecoration(
-                        borderRadius: BorderRadius.only(
-                            bottomLeft: Radius.circular(9),
-                            topLeft: Radius.circular(9)),
-                        border: Border.all(color: Colors.blue),
-                        color: isFavorite == false
-                            ? Colors.blue
-                            : Colors.transparent),
-                    child: Icon(Icons.messenger_outline),
-                  ),
-                ),
-              ),
-              Flexible(
-                flex: 1,
-                child: GestureDetector(
-                  onTap: () {
-                    setState(() {
-                      isFavorite = true;
-                    });
-                  },
-                  child: Container(
-                    width: double.infinity,
-                    decoration: BoxDecoration(
-                        borderRadius: BorderRadius.only(
-                            topRight: Radius.circular(9),
-                            bottomRight: Radius.circular(9)),
-                        border: Border.all(color: Colors.blue),
-                        color: isFavorite == false
-                            ? Colors.transparent
-                            : Colors.blue),
-                    child: Icon(Icons.favorite),
-                  ),
-                ),
-              ),
-            ],
-          ),
-        ));
-  }
-
-  @override
-  Widget body() {
-    return isFavorite
-        ? MyFavorites()
-        : Container(
-            width: SizeConfig.blockSizeHorizontal * 100,
-            height: SizeConfig.blockSizeVertical * 100,
-            child: Stack(
-              children: [
-                Center(
-                  child: Text(
-                    "Ingen beskeder",
-                    style: TextStyle(color: Colors.grey, fontSize: 30),
-                  ),
-                ),
-                Positioned(
-                  bottom: 100,
-                  child: Container(
-                    width: SizeConfig.blockSizeHorizontal * 100,
-                    child: Padding(
-                      padding: const EdgeInsets.symmetric(horizontal: 80),
-                      child: SizedBox(
-                        height: 50,
-                        child: RaisedButton(
-                          shape: RoundedRectangleBorder(
-                            borderRadius: BorderRadius.circular(30),
-                          ),
-                          child: Text(
-                            "Nyt Event",
-                            style: TextStyle(
-                              color: Colors.white,
-                              fontSize: 20,
-                            ),
-                          ),
-                          color: blue,
-                          onPressed: () {
-                            Navigator.push(
-                              this.context,
-                              FadeRoute(
-                                page: AddOrRepostEvent(),
-                              ),
-                            );
-                          },
-                        ),
-                      ),
-                    ),
-                  ),
-                ),
-              ],
-            ),
-          );
   }
 }
